chore: hoist dependencies into workspace tables
