+++ Cargo.toml
@@ -2,6 +2,99 @@
 resolver = "2"
 members = ["crates/*"]
 
+[workspace.package]
+version = "2.0.2"
+edition = "2024"
+rust-version = "1.85"
+authors = ["Kevin Delfour"]
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/systm-d/repolens"
+homepage = "https://github.com/systm-d/repolens"
+
+[workspace.dependencies]
+# Serialization
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+serde_yaml = "0.9"
+toml = "0.8"
+
+# Filesystem + globbing
+walkdir = "2"
+ignore = "0.4"
+globset = "0.4"
+
+# Text / patterns
+regex = "1.12"
+lazy_static = "1"
+url = "2"
+sha2 = "0.10"
+similar = "2"
+
+# Paths
+dirs = "5"
+
+# GitHub / HTTP
+octocrab = "0.41"
+which = "6"
+reqwest = { version = "0.12", features = ["json", "rustls-tls"], default-features = false }
+
+# Templating
+tera = "1"
+minijinja = "2"
+
+# Validation
+jsonschema = "0.29"
+
+# Parallelism / async
+rayon = "1.8"
+tokio = { version = "1", features = ["full"] }
+async-trait = "0.1"
+
+# Date/time
+chrono = { version = "0.4", features = ["serde"] }
+
+# Error handling
+anyhow = "1"
+thiserror = "1"
+
+# Logging / metrics
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
+metrics = "0.22"
+metrics-prometheus = "0.3"
+
+# Terminal output
+colored = "2"
+indicatif = { version = "0.17", features = ["tokio"] }
+console = "0.15"
+dialoguer = { version = "0.11", features = ["fuzzy-select"] }
+
+# CLI framework
+clap = { version = "4", features = ["derive"] }
+clap_mangen = "0.2"
+clap_complete = "4"
+clap_complete_nushell = "0.1"
+
+# Dev dependencies
+tempfile = "3"
+pretty_assertions = "1"
+criterion = { version = "0.5", features = ["html_reports"] }
+futures = "0.3"
+serial_test = "3.3.1"
+insta = "1"
+assert_cmd = "2"
+predicates = "3"
+
+[workspace.lints.rust]
+# Deliberate deviation from the template's `forbid`: `deny` allows the two
+# documented `#[allow(unsafe_code)]` env-var sites in repolens-core to compile.
+# See .superpowers/sdd/task-3-report.md.
+unsafe_code = "deny"
+unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tarpaulin)'] }
+
+[workspace.lints.clippy]
+all = { level = "warn", priority = -1 }
+
 [profile.release]
 lto = true
 codegen-units = 1
