+++ crates/repolens/Cargo.toml
@@ -1,100 +1,88 @@
 [package]
 name = "repolens"
-version = "2.0.2"
-edition = "2024"
-rust-version = "1.85"
-authors = ["Kevin Delfour"]
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
 description = "A CLI tool to audit and prepare repositories for open source or enterprise standards"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/systm-d/repolens"
-homepage = "https://github.com/systm-d/repolens"
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
 documentation = "https://docs.rs/repolens"
 keywords = ["cli", "audit", "github", "opensource", "security"]
 categories = ["command-line-utilities", "development-tools"]
 readme = "../../README.md"
 exclude = [
-    ".gitignore",
-    ".repolens.toml",
-    ".claude/",
-    "CLAUDE.md",
-    "DEVELOPMENT.md",
-    "RELEASE.md",
-    "TEST.md",
-    "TODO.md",
-    "deny.toml",
     "tests/",
     "benches/",
 ]
 
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = ['cfg(tarpaulin)'] }
+[lints]
+workspace = true
 
 [dependencies]
 # Core auditing library
 repolens-core = { path = "../repolens-core", version = "2.0.2" }
 
 # CLI framework
-clap = { version = "4", features = ["derive", "env", "color"] }
+clap = { workspace = true, features = ["env", "color"] }
 
 # Man page generation
-clap_mangen = "0.2"
+clap_mangen.workspace = true
 
 # Shell completion generation
-clap_complete = "4"
-clap_complete_nushell = "0.1"
+clap_complete.workspace = true
+clap_complete_nushell.workspace = true
 
 # Async runtime
-tokio = { version = "1", features = ["full"] }
+tokio.workspace = true
 
 # Serialization
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
+serde.workspace = true
+serde_json.workspace = true
 
 # JSON Schema validation (report validation in output layer)
-jsonschema = "0.29"
+jsonschema.workspace = true
 
 # Text diff for interactive mode
-similar = "2"
+similar.workspace = true
 
 # Terminal output
-colored = "2"
-indicatif = { version = "0.17", features = ["tokio"] }
-console = "0.15"
+colored.workspace = true
+indicatif.workspace = true
+console.workspace = true
 
 # Interactive prompts
-dialoguer = { version = "0.11", features = ["fuzzy-select"] }
+dialoguer.workspace = true
 
 # Error handling
-anyhow = "1"
-thiserror = "1"
+anyhow.workspace = true
+thiserror.workspace = true
 
 # Home directory resolution
-dirs = "5"
+dirs.workspace = true
 
 # Logging
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
+tracing.workspace = true
+tracing-subscriber.workspace = true
 
 # Date/time for reports
-chrono = { version = "0.4", features = ["serde"] }
+chrono.workspace = true
 
 [build-dependencies]
-clap_mangen = "0.2"
-clap = { version = "4", features = ["derive"] }
-clap_complete = "4"
-clap_complete_nushell = "0.1"
+clap_mangen.workspace = true
+clap.workspace = true
+clap_complete.workspace = true
+clap_complete_nushell.workspace = true
 
 [dev-dependencies]
-tempfile = "3"
-assert_cmd = "2"
-predicates = "3"
-pretty_assertions = "1"
-serial_test = "3.3.1"
-insta = "1"
-# TOML round-trip tests for HooksConfig (inline hooks tests)
-toml = "0.8"
+tempfile.workspace = true
+assert_cmd.workspace = true
+predicates.workspace = true
+pretty_assertions.workspace = true
+serial_test.workspace = true
+insta.workspace = true
 
 [[bin]]
 name = "repolens"
 path = "src/main.rs"
-
